create genre Genre entity
