+++ data/repository/src/main/java/com/berlin/repository/datasource/remote/dto/GenreDto.kt
@@ -1,0 +1,13 @@
+package com.berlin.repository.datasource.remote.dto
+
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class GenreDto(
+    @SerialName("id")
+    val id: Int?,
+    @SerialName("name")
+    val name: String?
+)
