+++ data/repository/src/main/java/com/berlin/repository/datasource/remote/dto/GenreResponse.kt
@@ -1,13 +1,9 @@
 package com.berlin.repository.datasource.remote.dto
 
-
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Genre(
-    @SerialName("id")
-    val id: Int?,
-    @SerialName("name")
-    val name: String?
+data class GenreResponse(
+    @SerialName("genres") val genres: List<GenreDto>
 )
