+++ domain/entity/src/main/java/com/berlin/entity/Genre.kt
@@ -1,0 +1,6 @@
+package com.berlin.entity
+
+data class Genre(
+    val id: Int?,
+    val name: String?
+)
